fix: revert directory rename
