+++ opensourceleg/control/fsm.py
@@ -20,7 +20,7 @@ from collections.abc import Iterator
 from typing import Any, Callable, Optional
 
 from opensourceleg.logging.logger import LOGGER
-from opensourceleg.timing import SoftRealtimeLoop
+from opensourceleg.time import SoftRealtimeLoop
 
 
 class State:
