+++ tutorials/actuators/dephy/current_control.py
@@ -1,7 +1,7 @@
 from opensourceleg.actuators.base import CONTROL_MODES
 from opensourceleg.actuators.dephy import DephyActuator
 from opensourceleg.logging.logger import Logger
-from opensourceleg.timing import SoftRealtimeLoop
+from opensourceleg.time import SoftRealtimeLoop
 
 TIME_TO_STEP = 1.0
 FREQUENCY = 1000
