+++ tutorials/actuators/dephy/voltage_control.py
@@ -3,7 +3,7 @@ import time
 from opensourceleg.actuators.base import CONTROL_MODES
 from opensourceleg.actuators.dephy import DephyActuator
 from opensourceleg.logging.logger import Logger
-from opensourceleg.timing import SoftRealtimeLoop
+from opensourceleg.time import SoftRealtimeLoop
 
 FREQUENCY = 1000
 TIME_TO_STEP = 1.0
