+++ tutorials/actuators/moteus/torque_control.py
@@ -5,7 +5,7 @@ from moteus import Register
 
 from opensourceleg.actuators.moteus import MoteusActuator
 from opensourceleg.logging.logger import LOGGER
-from opensourceleg.timing import SoftRealtimeLoop
+from opensourceleg.time import SoftRealtimeLoop
 
 TIME_TO_STEP = 1.0
 FREQUENCY = 200
